DependenciesParser.jl

DependenciesParser.jl

DependenciesParser.jl is a package for quickly obtaining the list of dependencies for packages.

DependenciesParser.installable — Function.

installable(pkg::AbstractString,
            julia::VersionNumber = VERSION;
            direct::Bool = false)::Tuple{Bool,Vector{String}}

Return whether the package is installable and the dependencies for the solved version.
If direct, only direct dependencies are returned.

source